Clarification for surface capabilities with extent matching the special value 0xFFFFFFFF (refs #90)

+++ base/vulkanswapchain.hpp
@@ -283,8 +283,8 @@ public:
 		assert(!err);
 
 		VkExtent2D swapchainExtent = {};
-		// width and height are either both -1, or both not -1.
-		if (surfCaps.currentExtent.width == -1)
+		// If width (and height) equals the special value 0xFFFFFFFF, the size of the surface will be set by the swapchain
+		if (surfCaps.currentExtent.width == (uint32_t)-1)
 		{
 			// If the surface size is undefined, the size is set to
 			// the size of the images requested.
